# core-banking-accounts-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-banking-accounts
  description: Banking Accounts Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/accounts
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/accounts/{accountId}/spaces/{accountSpaceId}/providers/{providerId}:
    get:
      tags:
        - Account Space Providers
      summary: Get Account Space Provider by ID
      description: Retrieve a specific provider record by its unique identifier, ensuring it belongs to the specified account space.
      operationId: getProviderForSpace
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account
          required: true
          schema:
            type: string
            format: uuid
        - name: accountSpaceId
          in: path
          description: Unique identifier of the account space
          required: true
          schema:
            type: string
            format: uuid
        - name: providerId
          in: path
          description: Unique identifier of the provider record
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved the account provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountProviderDTO'
        '404':
          description: Provider record not found
    put:
      tags:
        - Account Space Providers
      summary: Update Account Space Provider
      description: Update an existing provider record associated with the specified account space.
      operationId: updateProviderForSpace
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account
          required: true
          schema:
            type: string
            format: uuid
        - name: accountSpaceId
          in: path
          description: Unique identifier of the account space
          required: true
          schema:
            type: string
            format: uuid
        - name: providerId
          in: path
          description: Unique identifier of the provider record to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountProviderDTO'
        required: true
      responses:
        '200':
          description: Provider updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountProviderDTO'
        '404':
          description: Provider record not found
    delete:
      tags:
        - Account Space Providers
      summary: Delete Account Space Provider
      description: Remove an existing provider record from the specified account space by its unique identifier.
      operationId: deleteProviderForSpace
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account
          required: true
          schema:
            type: string
            format: uuid
        - name: accountSpaceId
          in: path
          description: Unique identifier of the account space
          required: true
          schema:
            type: string
            format: uuid
        - name: providerId
          in: path
          description: Unique identifier of the provider record to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Provider deleted successfully
        '404':
          description: Provider record not found
  /api/v1/accounts/{accountId}/providers/{providerId}:
    get:
      tags:
        - Account Providers
      summary: Get Account Provider by ID
      description: Retrieve a specific provider record by its unique identifier, ensuring it belongs to the specified account.
      operationId: getProvider
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account
          required: true
          schema:
            type: string
            format: uuid
        - name: providerId
          in: path
          description: Unique identifier of the provider record
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved the account provider
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountProviderDTO'
        '404':
          description: Provider record not found
    put:
      tags:
        - Account Providers
      summary: Update Account Provider
      description: Update an existing provider record associated with the specified account.
      operationId: updateProvider
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account
          required: true
          schema:
            type: string
            format: uuid
        - name: providerId
          in: path
          description: Unique identifier of the provider record to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountProviderDTO'
        required: true
      responses:
        '200':
          description: Provider updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountProviderDTO'
        '404':
          description: Provider record not found
    delete:
      tags:
        - Account Providers
      summary: Delete Account Provider
      description: Remove an existing provider record from the specified account by its unique identifier.
      operationId: deleteProvider
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account
          required: true
          schema:
            type: string
            format: uuid
        - name: providerId
          in: path
          description: Unique identifier of the provider record to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Provider deleted successfully
        '404':
          description: Provider record not found
...
  /api/v1/accounts/{accountId}:
    get:
      tags:
        - Accounts
      summary: Get Account by ID
      description: Retrieve an existing bank account by its unique identifier.
      operationId: getAccount
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDTO'
    put:
      tags:
        - Accounts
      summary: Update Account
      description: Update the details of an existing bank account by its unique identifier.
      operationId: updateAccount
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDTO'
    delete:
      tags:
        - Accounts
      summary: Delete Account
      description: Delete an existing bank account by its unique identifier.
      operationId: deleteAccount
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/account-spaces/{accountSpaceId}:
    get:
      tags:
        - Account Spaces
      summary: Get Account Space by ID
      description: Retrieve an existing account space by its unique identifier.
      operationId: getAccountSpace
      parameters:
        - name: accountSpaceId
          in: path
          description: Unique identifier of the account space
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSpaceDTO'
    put:
      tags:
        - Account Spaces
      summary: Update Account Space
      description: Update the details of an existing account space by its unique identifier.
      operationId: updateAccountSpace
      parameters:
        - name: accountSpaceId
          in: path
          description: Unique identifier of the account space to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountSpaceDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSpaceDTO'
    delete:
      tags:
        - Account Spaces
      summary: Delete Account Space
      description: Delete an existing account space by its unique identifier. Main spaces cannot be deleted.
      operationId: deleteAccountSpace
      parameters:
        - name: accountSpaceId
          in: path
          description: Unique identifier of the account space to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/accounts/{accountId}/spaces/{accountSpaceId}/providers:
    get:
      tags:
        - Account Space Providers
      summary: List Account Space Providers
      description: Retrieve a paginated list of provider records associated with the specified account space.
      operationId: getAllProvidersForSpace
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account
          required: true
          schema:
            type: string
            format: uuid
        - name: accountSpaceId
          in: path
          description: Unique identifier of the account space
          required: true
          schema:
            type: string
            format: uuid
        - name: pageNumber
          in: query
          description: The zero-based page number to retrieve.
          required: false
          schema:
            type: string
            description: The zero-based page number to retrieve.
            example: 0
            default: '0'
          example: 0
        - name: pageSize
          in: query
          description: The number of items per page.
          required: false
          schema:
            type: string
            description: The number of items per page.
            example: 10
            default: '10'
          example: 10
        - name: sortBy
          in: query
          description: The field to sort the results by.
          required: false
          schema:
            type: string
            description: The field to sort the results by.
            example: name
          example: name
        - name: sortDirection
          in: query
          description: The direction of sorting, either ASC or DESC.
          required: false
          schema:
            type: string
            description: The direction of sorting, either ASC or DESC.
            example: DESC
            default: DESC
          example: DESC
      responses:
        '200':
          description: Successfully retrieved the provider records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '404':
          description: No providers found for the specified account space
    post:
      tags:
        - Account Space Providers
      summary: Create Account Space Provider
      description: Create a new provider record and associate it with the specified account space.
      operationId: createProviderForSpace
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account
          required: true
          schema:
            type: string
            format: uuid
        - name: accountSpaceId
          in: path
          description: Unique identifier of the account space
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountProviderDTO'
        required: true
      responses:
        '201':
          description: Provider created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountProviderDTO'
        '400':
          description: Invalid provider data provided
...
  /api/v1/account-restrictions/account/{accountId}:
    get:
      tags:
        - Account Restrictions
      summary: Get Account Restrictions by Account ID
      description: Retrieve all restrictions for a specific account.
      operationId: getAccountRestrictionsByAccountId
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved the restrictions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountRestrictionDTO'
        '400':
          description: Invalid account ID
...
  /api/v1/account-notifications/account/{accountId}:
    get:
      tags:
        - Account Notifications
      summary: Get Account Notifications by Account ID
      description: Retrieve all notifications for a specific account.
      operationId: getAccountNotificationsByAccountId
      parameters:
        - name: accountId
          in: path
          description: Unique identifier of the account
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved the notifications
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountNotificationDTO'
        '400':
          description: Invalid account ID
...
components:
  schemas:
...
    AccountProviderDTO:
      required:
        - accountId
        - providerName
        - status
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        accountProviderId:
          type: string
          format: uuid
          readOnly: true
        accountId:
          type: string
          format: uuid
        providerName:
          maxLength: 100
          minLength: 0
          type: string
        externalReference:
          maxLength: 100
          minLength: 0
          type: string
        status:
          type: string
          enum:
            - ACTIVE
            - INACTIVE
            - PENDING
            - SUSPENDED
        accountSpaceId:
          type: string
          format: uuid
...
    AccountDTO:
      required:
        - accountNumber
        - accountStatus
        - accountType
        - branchId
        - contractId
        - currency
        - openDate
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        accountId:
          type: string
          format: uuid
          readOnly: true
        contractId:
          type: string
          format: uuid
        accountNumber:
          maxLength: 50
          minLength: 0
          type: string
        accountType:
          type: string
          enum:
            - CHECKING
            - SAVINGS
            - TERM_DEPOSIT
            - LOAN
            - CREDIT_CARD
            - INVESTMENT
            - MORTGAGE
            - BUSINESS_CHECKING
            - BUSINESS_SAVINGS
            - FOREIGN_CURRENCY
            - PREMIUM
            - YOUTH
            - SENIOR
            - PENSION
            - GOVERNMENT_BENEFITS
            - CRYPTO_WALLET
            - TOKENIZED_ASSET
            - DIGITAL_ASSET
            - STABLECOIN
            - NFT_COLLECTION
        currency:
          maxLength: 3
          minLength: 3
          pattern: ^[A-Z]{3}$
          type: string
        openDate:
          type: string
          format: date
        closeDate:
          type: string
          format: date
        accountStatus:
          type: string
          enum:
            - OPEN
            - CLOSED
            - SUSPENDED
            - DORMANT
        branchId:
          type: string
          format: uuid
        description:
          maxLength: 500
          minLength: 0
          type: string
        accountSubType:
          type: string
          enum:
            - STANDARD_CHECKING
            - INTEREST_CHECKING
            - REWARDS_CHECKING
            - FREE_CHECKING
            - PREMIUM_CHECKING
            - STANDARD_SAVINGS
            - HIGH_YIELD_SAVINGS
            - MONEY_MARKET
            - CERTIFICATE_OF_DEPOSIT
            - IRA
            - FIXED_TERM_DEPOSIT
            - VARIABLE_TERM_DEPOSIT
            - PERSONAL_LOAN
            - AUTO_LOAN
            - HELOC
            - STANDARD_CREDIT_CARD
            - REWARDS_CREDIT_CARD
            - CASHBACK_CREDIT_CARD
            - TRAVEL_CREDIT_CARD
            - BUSINESS_CREDIT_CARD
            - FIXED_RATE_MORTGAGE
            - ADJUSTABLE_RATE_MORTGAGE
            - INTEREST_ONLY_MORTGAGE
            - STANDARD_BUSINESS_CHECKING
            - PREMIUM_BUSINESS_CHECKING
            - STANDARD_BUSINESS_SAVINGS
            - PREMIUM_BUSINESS_SAVINGS
            - MULTI_CURRENCY
            - SINGLE_CURRENCY_FOREIGN
            - RELATIONSHIP_PREMIUM
            - WEALTH_MANAGEMENT
            - YOUTH_SAVINGS
            - STUDENT_CHECKING
            - SENIOR_CHECKING
            - SENIOR_SAVINGS
            - PENSION_DISTRIBUTION
            - SOCIAL_SECURITY_BENEFITS
            - GOVERNMENT_ASSISTANCE
        taxReportingStatus:
          type: string
          enum:
            - REPORTABLE
            - EXEMPT
            - BACKUP_WITHHOLDING
            - FATCA_REPORTABLE
            - CRS_REPORTABLE
            - FATCA_CRS_REPORTABLE
            - PENDING_DOCUMENTATION
            - INCOMPLETE_INFORMATION
            - NON_REPORTABLE
        regulatoryStatus:
          type: string
          enum:
            - COMPLIANT
            - PENDING_DOCUMENTATION
            - ENHANCED_DUE_DILIGENCE
            - UNDER_REVIEW
            - SPECIAL_MONITORING
            - SANCTIONS_SCREENING
            - PEP_MONITORING
            - HIGH_RISK_MONITORING
            - REGULATORY_EXEMPTION
            - NON_COMPLIANT
            - REGULATORY_RESTRICTED
        maturityDate:
          type: string
          format: date
        interestAccrualMethod:
          type: string
          enum:
            - SIMPLE
            - DAILY_COMPOUND
            - MONTHLY_COMPOUND
            - QUARTERLY_COMPOUND
            - SEMI_ANNUAL_COMPOUND
            - ANNUAL_COMPOUND
            - CONTINUOUS_COMPOUND
            - DAILY_AVERAGE_BALANCE
            - MONTHLY_AVERAGE_BALANCE
            - MINIMUM_BALANCE
            - TIERED_RATE
            - STEPPED_RATE
            - DAY_COUNT_30_360
            - DAY_COUNT_ACTUAL_360
            - DAY_COUNT_ACTUAL_365
            - DAY_COUNT_ACTUAL_ACTUAL
        interestPaymentFrequency:
          type: string
          enum:
            - DAILY
            - WEEKLY
            - BI_WEEKLY
            - MONTHLY
            - QUARTERLY
            - SEMI_ANNUALLY
            - ANNUALLY
            - AT_MATURITY
            - AT_CLOSURE
            - SPECIFIC_DAY_OF_MONTH
            - SPECIFIC_DATE
            - NONE
        minimumBalance:
          minimum: 0
          exclusiveMinimum: false
          type: number
        overdraftLimit:
          minimum: 0
          exclusiveMinimum: false
          type: number
        walletAddress:
          maxLength: 100
          minLength: 0
          type: string
        blockchainNetwork:
          maxLength: 50
          minLength: 0
          type: string
        tokenContractAddress:
          maxLength: 100
          minLength: 0
          type: string
        tokenStandard:
          maxLength: 20
          minLength: 0
          type: string
        isCustodial:
          type: boolean
      description: Data for the new account
    AccountSpaceDTO:
      required:
        - accountId
        - balance
        - spaceName
        - spaceType
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        accountSpaceId:
          type: string
          format: uuid
          readOnly: true
        accountId:
          type: string
          format: uuid
        spaceName:
          maxLength: 100
          minLength: 0
          type: string
        spaceType:
          type: string
          enum:
            - MAIN
            - SAVINGS
            - VACATION
            - EMERGENCY
            - GOALS
            - CUSTOM
        balance:
          minimum: 0
          exclusiveMinimum: false
          type: number
        targetAmount:
          minimum: 0
          exclusiveMinimum: true
          type: number
        targetDate:
          type: string
          format: date-time
        iconId:
          maxLength: 50
          minLength: 0
          type: string
        colorCode:
          maxLength: 7
          minLength: 0
          pattern: ^#[0-9A-Fa-f]{6}$
          type: string
        description:
          maxLength: 500
          minLength: 0
          type: string
        isVisible:
          type: boolean
        goalProgressPercentage:
          type: number
          readOnly: true
        remainingToTarget:
          type: number
          readOnly: true
        estimatedCompletionDate:
          type: string
          format: date-time
          readOnly: true
        isGoalCompleted:
          type: boolean
          readOnly: true
        enableAutomaticTransfers:
          type: boolean
        transferFrequency:
          type: string
          enum:
            - DAILY
            - WEEKLY
            - MONTHLY
            - QUARTERLY
            - ANNUALLY
        transferAmount:
          type: number
        sourceSpaceId:
          type: string
          format: uuid
        averageGrowthRate:
          type: number
          readOnly: true
        contributionToAccountTotal:
          type: number
          readOnly: true
        isFrozen:
          type: boolean
        frozenDateTime:
          type: string
          format: date-time
          readOnly: true
        unfrozenDateTime:
          type: string
          format: date-time
          readOnly: true
        lastBalanceUpdateReason:
          type: string
          readOnly: true
        lastBalanceUpdateDateTime:
          type: string
          format: date-time
          readOnly: true
      description: Data for the new account space
...